trending refactor
